[CHE-8] Move Klarna data models into their own file
